UI: repeat= — a template over a list, as a fixed pool with a moving window
The tree is expanded once, at load, and never changes shape again. That is
the design rather than a shortcut: UIElement::structureEpoch() is
process-wide, so a list that added and removed children would make every
binder in EVERY document re-collect and re-resolve on the frames it changed
size. Here the elements stand still and the data slides through them, and
because every UIDataSource setter is equality-gated, a slot whose row did
not change writes nothing and re-applies nothing. Sliding a window costs
exactly the slots that changed; an idle list costs three integer compares.

    <Element repeat="inventory" repeat-count="5" repeat-offset="invScroll">
      <Element class="bag-row">
        <Label text="{$index}"/>
        <Label text="{name}"/>      <!-- bare: reads the ROW -->
        <Label text="x{count}"/>
      </Element>
    </Element>

UIRecord and UIList are the data. A list is deliberately NOT a property: no
{hole} can read one, it has its own version, and SetList therefore does not
move the whole-source stamp — waking every health and score binding for a
change none of them can see would be pure cost.

Five things the design review made non-negotiable, each of which had a
concrete failure behind it:

Every column the template reads is SEEDED on every slot at build time. A
slot the window never reaches receives no row, and UIDataSource has no way
to declare a property other than writing one — so without this an unreached
slot's bindings never resolve, unresolvedCount() stays non-zero forever, and
UpdateToTarget's pending-retry check turns every frame in which any source
moves into a full document Rebuild, two or three times per document.

The engine's own columns are $slot/$index/$count/$present. findOrAdd_
resolves by name alone, so a row column called `index` and an engine-written
`index` would be the SAME property: they would alternate every frame, bump
the version every frame, and pin that slot hot for the life of the process.
A '$' cannot collide, and the hole grammar already accepts it.

Teardown unregisters from the names it actually registered. Deriving them
from the new spec list, a count delta or the list name all fail identically
the moment an author DELETES the repeat — and a stale registration is not a
stale read, it is a freed UIDataSource* that sourceVersionSum() dereferences
on the next frame.

A template root carrying if=, data-source= or repeat= is a load error, not a
silent skip. Losing the row's visibility binding leaves surplus rows in
layout, painted and hit-testable, which is the exact opposite of what the
author asked for.

Slot sources are registered but hidden from sourceNames(), and a bare hole
that no row provides is checked once against the data rather than once per
clone. One typo answered with 64 identical lines, each listing 64 generated
names, is not a diagnostic.

Also widens UIBinder::Entry's four pool offsets from 16 to 32 bits: resolve_
assigns pool sizes into them without a range check, and every bounds check
downstream catches an index that is out of range rather than one that
wrapped — a wrapped index is in range and silently reads another hole.

52 tests in test_ui_repeat.cpp, two of them through the shipped HUD and
UIWorld, which is the only path that runs UpdateRepeats where a game does.
The sample HUD grows a 12-row inventory behind a 5-slot pool with PREV/NEXT.

+++ Engine/src/ui/UIBinding.h
@@ -202,9 +202,15 @@ namespace MyCoreEngine::ui {
             UIDataSource*    src = nullptr;      // null => pending resolution
             std::uint32_t    lastSourceVersion = 0;
             std::int32_t     propIndex = -1;
-            std::uint16_t    convFirst = 0, convCount = 0;
+            // 32-bit, not 16: resolve_ assigns pool sizes into these without a
+            // range check, and the bounds checks downstream catch an index that
+            // is OUT of range, not one that WRAPPED — a wrapped index is in
+            // range and silently reads somebody else's hole. A `repeat=` of 64
+            // slots multiplies every template in its subtree by 64, which is
+            // how a document gets within reach of 65536 in the first place.
+            std::uint32_t    convFirst = 0, convCount = 0;
             // Range into holePool_, parallel to convFirst/convCount.
-            std::uint16_t    holeFirst = 0, holeCount = 0;
+            std::uint32_t    holeFirst = 0, holeCount = 0;
             bool             layoutAffecting = false;
             bool             reported = false;   // latch, re-armed on a kind change
             std::uint8_t     reportedKind = 0;
